Time: 217 ms (82.60%), Space: 102.4 MB (40.31%) - LeetHub

+++ 916-word-subsets/916-word-subsets.cpp
@@ -11,8 +11,8 @@ public:
             vector<int> hashTemp(26, 0);
             for(int j = 0; j < s.size(); j++)
             {
-                hashTemp[(int)(s[j]-'a')]++;
-                hash[(int)(s[j]-'a')] = max(hashTemp[(int)(s[j]-'a')], hash[(int)(s[j]-'a')]);
+                hashTemp[s[j]-'a']++;
+                hash[s[j]-'a'] = max(hashTemp[s[j]-'a'], hash[s[j]-'a']);
             }
         }
         
@@ -23,7 +23,7 @@ public:
             bool flag = true;
             
             for(int j = 0; j < s.size(); j++)
-                hashTemp[(int)(s[j]-'a')]++;
+                hashTemp[s[j]-'a']++;
            
             for(int j = 0; j < 26; j++)
                 if(hashTemp[j] < hash[j]) {flag = false; break;}
